refactor(landing): extract shared LegalPage template
TermsPage and PrivacyPage were 95% identical — same layout, same
header, same intro paragraph, same section-mapping loop. The only
differences were:

  1. useTranslations namespace ("legal.terms" vs "legal.privacy")
  2. testid ("legal-terms-page" vs "legal-privacy-page")
  3. SECTION_KEYS array contents (10 vs 9 items)

Pulled the duplicated JSX into a new `LegalPage` template that takes
namespace, testId, and sectionKeys as props. TermsPage and PrivacyPage
now declare their section list and delegate rendering.

Net: one source of truth for legal-page structure. Any future styling
or layout change happens in one place.

No test changes required — existing tests assert rendered output
(testid presence, h1 presence, last-updated, section heading count),
all of which are unchanged.
